requirements.txt
aiohttp==3.6.2
arrow==0.15.6
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
blessed==1.17.8
boto3==1.14.4
botocore==1.17.4
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
Django==3.0.7
django-background-tasks==1.2.5
django-compat==1.0.15
django-filter==2.3.0
django-picklefield==3.0.1
django-q==1.2.4
django-storages==1.9.1
djangorestframework==3.11.0
docopt==0.6.2
docutils==0.15.2
et-xmlfile==1.0.1
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
Markdown==3.2.2
MarkupSafe==1.1.1
multidict==4.7.6
openpyxl==3.0.3
Pillow==7.1.2
pyee==7.0.2
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
redis==3.5.3
requests==2.23.0
res==0.1.7
s3transfer==0.3.3
six==1.15.0
slackclient==2.7.1
slackeventsapi==2.1.0
sqlparse==0.3.1
urllib3==1.25.9
wcwidth==0.2.4
Werkzeug==1.0.1
wget==3.2
yarl==1.4.2
